+++ bin/install_material_symbols_icons_fonts.dart
@@ -45,34 +45,37 @@ String getRootPathsToLatestInstalledPackage() {
   rootDir = path.dirname(pathToScript);
 
   // following for testing
-  if(!rootDir.contains('global_packages')) {
-    rootDir = r"C:\Users\Tim\AppData\Local\Pub\Cache\global_packages\dart_frog_cli\bin";
+  if (!rootDir.contains('global_packages')) {
+    rootDir =
+        r"C:\Users\Tim\AppData\Local\Pub\Cache\global_packages\dart_frog_cli\bin";
   }
 
-  String pubDevPackagesDir = path.join(rootDir, '..', '..', '..', 'hosted', 'pub.dev');
+  String pubDevPackagesDir =
+      path.join(rootDir, '..', '..', '..', 'hosted', 'pub.dev');
 
-  if(debugScripts) print('pubDevPackagesDir=$pubDevPackagesDir');
+  if (debugScripts) print('pubDevPackagesDir=$pubDevPackagesDir');
 
-  final packageDirs = Glob('material_symbols_icons-*', caseSensitive:false, recursive:false);
+  final packageDirs =
+      Glob('material_symbols_icons-*', caseSensitive: false, recursive: false);
   final baseToChop = 'material_symbols_icons-';
 
-  final listFSE = packageDirs.listSync(root:pubDevPackagesDir);
+  final listFSE = packageDirs.listSync(root: pubDevPackagesDir);
   String highestVersion = '4.2600.0';
   String latestPackageDir = '';
 
-  for(final fse in listFSE) {
+  for (final fse in listFSE) {
     String dirName = fse.basename;
     String version = dirName.substring(baseToChop.length);
-    if(debugScripts) print('Found directory $dirName version=$version');
-    if(version.length>=8) {
-      if(version.compareTo(highestVersion)>0) {
+    if (debugScripts) print('Found directory $dirName version=$version');
+    if (version.length >= 8) {
+      if (version.compareTo(highestVersion) > 0) {
         highestVersion = version;
         latestPackageDir = fse.path;
       }
     }
   }
-  if(debugScripts) print('Highest Version = $highestVersion');
-  if(debugScripts) print('latestPackageDir = $latestPackageDir');
+  if (debugScripts) print('Highest Version = $highestVersion');
+  if (debugScripts) print('latestPackageDir = $latestPackageDir');
   return path.join(latestPackageDir, 'bin');
 }
 
@@ -103,8 +106,7 @@ void main(List<String> args) async {
       Options.debug.name,
       defaultsTo: false,
       negatable: false,
-      help:
-          'Debug scripts.',
+      help: 'Debug scripts.',
     )
     //OBSOLETE//..addOption(
     //OBSOLETE//  Options.path.name,
@@ -124,12 +126,12 @@ void main(List<String> args) async {
       negatable: false,
       help:
           'MacOS specific flag to additionally validate fonts using FontBook.',
-    )..addFlag(
+    )
+    ..addFlag(
       Options.uninstall.name,
       defaultsTo: false,
       negatable: false,
-      help:
-          'Uninstall the material symbols icons fonts.',
+      help: 'Uninstall the material symbols icons fonts.',
     );
 
   late final ArgResults parsedArgs;
@@ -192,9 +194,11 @@ void installMaterialSymbolsIconsFonts() {
       installMaterialSymbolsIconsFontLinux();
       break;
     default:
-      print(chalk.redBright('Unsupported operating system: ${Platform.operatingSystem}'));
+      print(chalk.redBright(
+          'Unsupported operating system: ${Platform.operatingSystem}'));
   }
 }
+
 void uninstallMaterialSymbolsIconsFonts() {
   print(chalk.cyanBright('running on ${Platform.operatingSystem}'));
   switch (Platform.operatingSystem) {
@@ -202,64 +206,76 @@ void uninstallMaterialSymbolsIconsFonts() {
       uninstallMaterialSymbolsIconsFontWindows();
       break;
     case 'macos':
-      print(chalk.redBright('Uninstalling Material Symbols Icons fonts is not supported on MacOS yet.'));
+      print(chalk.redBright(
+          'Uninstalling Material Symbols Icons fonts is not supported on MacOS yet.'));
       //uninstallMaterialSymbolsIconsFontMacOS();
       break;
     case 'linux':
-      print(chalk.redBright('Uninstalling Material Symbols Icons fonts is not supported on Linux yet.'));
+      print(chalk.redBright(
+          'Uninstalling Material Symbols Icons fonts is not supported on Linux yet.'));
       //uninstallMaterialSymbolsIconsFontLinux();
       break;
     default:
-      print(chalk.redBright('Unsupported operating system: ${Platform.operatingSystem}'));
+      print(chalk.redBright(
+          'Unsupported operating system: ${Platform.operatingSystem}'));
   }
 }
 
 //Windows specific functions
 void installMaterialSymbolsIconsFontWindows() {
-  print(chalk.cyanBright('running powershell scripts to install Material Symbols Icons fonts...'));
+  print(chalk.cyanBright(
+      'running powershell scripts to install Material Symbols Icons fonts...'));
   runPowerShellInstallFont(r'..\lib\fonts\MaterialSymbolsOutlined.ttf');
   runPowerShellInstallFont(r'..\lib\fonts\MaterialSymbolsRounded.ttf');
   runPowerShellInstallFont(r'..\lib\fonts\MaterialSymbolsSharp.ttf');
 }
 
 void uninstallMaterialSymbolsIconsFontWindows() {
-  print(chalk.cyanBright('running powershell scripts to UNINSTALL Material Symbols Icons fonts...'));
+  print(chalk.cyanBright(
+      'running powershell scripts to UNINSTALL Material Symbols Icons fonts...'));
   runPowerShellUninstallFont(r'..\lib\fonts\MaterialSymbolsOutlined.ttf');
   runPowerShellUninstallFont(r'..\lib\fonts\MaterialSymbolsRounded.ttf');
   runPowerShellUninstallFont(r'..\lib\fonts\MaterialSymbolsSharp.ttf');
 }
 
 void runPowerShellInstallFont(String fontNameWithRelativePath) {
-  var result = runPowerShellScriptOneArg(r'.\Install-Font.ps1', fontNameWithRelativePath);
-  final fontname = path.basename(path.withoutExtension(fontNameWithRelativePath));
+  var result = runPowerShellScriptOneArg(
+      r'.\Install-Font.ps1', fontNameWithRelativePath);
+  final fontname =
+      path.basename(path.withoutExtension(fontNameWithRelativePath));
   final numberFacesInstalled = int.tryParse(result);
-  if(numberFacesInstalled != null && numberFacesInstalled > 0) {
-    print(chalk.greenBright('$fontname font was successfully installed ($numberFacesInstalled faces installed).'));
+  if (numberFacesInstalled != null && numberFacesInstalled > 0) {
+    print(chalk.greenBright(
+        '$fontname font was successfully installed ($numberFacesInstalled faces installed).'));
   } else {
-    print(chalk.redBright('$fontname font was not installed likely because the font $fontNameWithRelativePath was not found.'));
+    print(chalk.redBright(
+        '$fontname font was not installed likely because the font $fontNameWithRelativePath was not found.'));
   }
 }
 
 void runPowerShellUninstallFont(String fontNameWithRelativePath) {
-  var result = runPowerShellScriptOneArg(r'.\Uninstall-Font.ps1', fontNameWithRelativePath);
-  final fontname = path.basename(path.withoutExtension(fontNameWithRelativePath));
-  if(result.toLowerCase().contains('True'.toLowerCase())) {
+  var result = runPowerShellScriptOneArg(
+      r'.\Uninstall-Font.ps1', fontNameWithRelativePath);
+  final fontname =
+      path.basename(path.withoutExtension(fontNameWithRelativePath));
+  if (result.toLowerCase().contains('True'.toLowerCase())) {
     print(chalk.greenBright('$fontname font was successfully uninstalled.'));
   } else {
-    print(chalk.redBright('$fontname font was not uninstalled because it was not currently installed.'));
+    print(chalk.redBright(
+        '$fontname font was not uninstalled because it was not currently installed.'));
   }
 }
 
 String runPowerShellScriptOneArg(String scriptPath, String argumentToScript) {
-  return runPowerShellScript(scriptPath, [argumentToScript] );
+  return runPowerShellScript(scriptPath, [argumentToScript]);
 }
 
 String runPowerShellScript(String scriptPath, List<String> argumentsToScript) {
-  final processResult = Process.runSync(
-          'Powershell.exe', ['-executionpolicy', 'bypass', '-File', scriptPath, ...argumentsToScript],
-          workingDirectory: rootDir //path.join(Directory.current.path, 'bin')
-        );
-  if(debugScripts) {
+  final processResult = Process.runSync('Powershell.exe',
+      ['-executionpolicy', 'bypass', '-File', scriptPath, ...argumentsToScript],
+      workingDirectory: rootDir //path.join(Directory.current.path, 'bin')
+      );
+  if (debugScripts) {
     print(chalk.yellowBright('Executing $scriptPath with $argumentsToScript'));
     print(chalk.redBright(processResult.stderr as String));
     print(chalk.blueBright(processResult.stdout as String));
@@ -269,17 +285,16 @@ String runPowerShellScript(String scriptPath, List<String> argumentsToScript) {
 
 void runShellInstallFontsScriptLinux() {
   String scriptName = 'install-fonts.sh';
-  if(macOSUseFontBook) {
+  if (macOSUseFontBook) {
     scriptName = 'install-fonts-withFontBook.sh';
   }
-  final scriptPath = path.join(rootDir, scriptName);//path.join('..', '..', 'bin', scriptName);
-  final fontWorkingDir = path.join(rootDir, '..', 'lib', 'fonts' );
-  //print(chalk.red('scriptPath=$scriptPath  fontWorkingDir=$fontWorkingDir'));  
-  final processResult = Process.runSync(
-          'sh', [ scriptPath ],
-          workingDirectory: fontWorkingDir
-        );
-  if(debugScripts) {
+  final scriptPath = path.join(
+      rootDir, scriptName); //path.join('..', '..', 'bin', scriptName);
+  final fontWorkingDir = path.join(rootDir, '..', 'lib', 'fonts');
+  //print(chalk.red('scriptPath=$scriptPath  fontWorkingDir=$fontWorkingDir'));
+  final processResult =
+      Process.runSync('sh', [scriptPath], workingDirectory: fontWorkingDir);
+  if (debugScripts) {
     print(chalk.yellowBright('Executed $scriptName'));
     print(chalk.redBright(processResult.stderr as String));
     print(chalk.blueBright(processResult.stdout as String));
@@ -288,17 +303,16 @@ void runShellInstallFontsScriptLinux() {
 
 void runShellInstallFontsScriptGloballyOnMacOS() {
   String scriptName = 'install-fonts-macAlt.sh';
-  if(macOSUseFontBook) {
+  if (macOSUseFontBook) {
     scriptName = 'install-fonts-macAlt-withFontBook.sh';
   }
-  final scriptPath = path.join(rootDir, scriptName);//path.join('..', '..', 'bin', scriptName);
-  final fontWorkingDir = path.join(rootDir, '..', 'lib', 'fonts' );
+  final scriptPath = path.join(
+      rootDir, scriptName); //path.join('..', '..', 'bin', scriptName);
+  final fontWorkingDir = path.join(rootDir, '..', 'lib', 'fonts');
   //print(chalk.red('scriptPath=$scriptPath  fontWorkingDir=$fontWorkingDir'));
-  final processResult = Process.runSync(
-          'sh', [ scriptPath ],
-          workingDirectory: fontWorkingDir
-        );
-  if(debugScripts) {
+  final processResult =
+      Process.runSync('sh', [scriptPath], workingDirectory: fontWorkingDir);
+  if (debugScripts) {
     print(chalk.yellowBright('Executed $scriptName'));
     print(chalk.redBright(processResult.stderr as String));
     print(chalk.blueBright(processResult.stdout as String));
@@ -307,25 +321,30 @@ void runShellInstallFontsScriptGloballyOnMacOS() {
 
 //MacOS specific functions
 void installMaterialSymbolsIconsFontMacOS() {
-  if(globalMacOSInstall) {
-    print(chalk.greenBright('Install Material Symbols Icons fonts globally${(macOSUseFontBook) ? ' and validating with FontBook.':'.'}'));
+  if (globalMacOSInstall) {
+    print(chalk.greenBright(
+        'Install Material Symbols Icons fonts globally${(macOSUseFontBook) ? ' and validating with FontBook.' : '.'}'));
     runShellInstallFontsScriptGloballyOnMacOS();
   } else {
-    print(chalk.greenBright('Install Material Symbols Icons fonts for current user${(macOSUseFontBook) ? ' and validating with FontBook.':'.'}...'));
+    print(chalk.greenBright(
+        'Install Material Symbols Icons fonts for current user${(macOSUseFontBook) ? ' and validating with FontBook.' : '.'}...'));
     runShellInstallFontsScriptLinux();
   }
 }
 
 void uninstallMaterialSymbolsIconsFontMacOS() {
-  print(chalk.redBright('UNINSTALLING Material Symbols Icons fonts not supported on MacOS.'));
+  print(chalk.redBright(
+      'UNINSTALLING Material Symbols Icons fonts not supported on MacOS.'));
 }
 
 //Linux specific functions
 void installMaterialSymbolsIconsFontLinux() {
-  print(chalk.greenBright('Install Material Symbols Icons fonts for current user using Linux script...'));
+  print(chalk.greenBright(
+      'Install Material Symbols Icons fonts for current user using Linux script...'));
   runShellInstallFontsScriptLinux();
 }
 
 void uninstallMaterialSymbolsIconsFontLinux() {
-  print(chalk.redBright('UNINSTALLING Material Symbols Icons fonts not supported on Linux.'));
+  print(chalk.redBright(
+      'UNINSTALLING Material Symbols Icons fonts not supported on Linux.'));
 }